feat(org): split components
