+++ src/apps/Organisations/index.jsx
@@ -117,259 +117,8 @@ const addMember = async (api, accountPair, id, target) => {
 	target.disabled = false
 }
 
-//
-//
-//
 
-const defaultContent = {}
 
-const Item = ({ content, mode }) => {
-	const { api } = useSubstrate()
-	const { address } = useWallet()
-	const navigate = useNavigate()
-
-	const [itemContent, setItemContent] = useState({})
-	const [metadata, setMetadata] = useState({})
-	const [imageURL, setImageURL] = useState(null)
-
-	const updateContent = (newContent) => setItemContent({ ...itemContent, ...newContent })
-
-	useEffect(() => {
-		if (!content) return
-		updateContent(content)
-	}, [content])
-
-	// get offchain data
-	useEffect(() => {
-		if (!itemContent.cid || itemContent.cid.length < 3) return
-		fetch(gateway + itemContent.cid)
-			.then((res) => res.text())
-			.then((txt) => setMetadata(JSON.parse(txt)))
-			.catch((err) => console.log(err))
-	}, [itemContent])
-
-	// get project image
-
-	useEffect(() => {
-		if (!metadata) return
-		setImageURL(metadata.logo ? gateway + metadata.logo : null)
-	}, [metadata])
-
-	// get onchain data
-
-	useEffect(() => {
-		if (!content) return
-		const query = async () => {
-			try {
-				const [
-					// config,
-					controller,
-					access,
-					memberState,
-					memberCount,
-					treasury,
-				] = await Promise.all([
-					// api.query.gameDaoControl.bodyConfig(content.id),
-					api.query.gameDaoControl.bodyController(content.id),
-					api.query.gameDaoControl.bodyAccess(content.id),
-					api.query.gameDaoControl.bodyMemberState([content.id, address]),
-					api.query.gameDaoControl.bodyMemberCount(content.id),
-					api.query.gameDaoControl.bodyTreasury(content.id),
-				])
-				updateContent({
-					// ...config.toHuman(),
-					controller: controller.toHuman(),
-					access: access.toHuman(),
-					memberState: memberState.toHuman(),
-					memberCount: memberCount.toHuman(),
-					treasury: treasury.toHuman(),
-				})
-			} catch (err) {
-				console.error(err)
-			}
-		}
-		query()
-	}, [api, content])
-
-	// useEffect(() => {
-
-	// 	if (!content.treasury) return
-
-	// 	const query = async () => {
-	// 		const unsub = await api.queryMulti([
-	// 			[api.query.system.account, content.treasury],
-	// 		], ([{ data: balance }]) => {
-	// 			console.log(`----> treasury of ${balance.free}`);
-	// 			setItemContent({
-	// 				...itemContent,
-	// 				treasuryBalance: balance.free
-	// 			})
-	// 		});
-	// 	}
-	// 	query()
-
-	// }, [content, api.query.system.account])
-
-	//
-	//
-	//
-
-	const handleMembership = async (e) => {
-		const op = e.target.value
-
-		switch (op) {
-			case '0':
-				console.log('join')
-				await addMember(api, accountPair, itemContent.id, e.target)
-				// join
-				return
-			case '1':
-				console.log('apply')
-				// apply for membership
-				return
-			case '2':
-				console.log('leave')
-				// remove member
-				return
-			default:
-				return
-		}
-	}
-
-	const handleAdmin = () => navigate('/app/organisations/admin/1234')
-	const handleDashboard = () => navigate('/app/organisations/dashboard/1234');
-
-	const buttonText = ['join', 'apply', 'leave']
-
-	const isAdmin = () => (address === itemContent.controller ? true : false)
-	const isMember = () => (itemContent.memberState > 0 ? true : false)
-
-	const Interactions = () => {
-		const text = buttonText[itemContent.access]
-		return (
-			<>
-				{(isMember() || isAdmin()) && (
-					<Button variant={'outlined'} fullWidth onClick={handleDashboard} value={itemContent.access}>{`Dashboard`}</Button>
-				)}
-				{isMember() && !isAdmin() && <Button variant={'outlined'} fullWidth onClick={handleMembership} value={itemContent.access}>{`leave`}</Button>}
-				{!isMember() && text && <Button variant={'outlined'} fullWidth onClick={handleMembership} value={itemContent.access}>{`${text}`}</Button>}
-				{isAdmin() && (
-					<Button variant={'outlined'} fullWidth onClick={handleAdmin}>
-						Admin
-					</Button>
-				)}
-			</>
-		)
-	}
-
-	const bodyToText = () => d.dao_bodies.filter((b) => b.value === Number(content.body))[0].text
-
-	const metaContent = React.useMemo(() => {
-		return (
-			<Stack sx={{ width: '100%', height: '100%' }} direction={'column'} justifyContent={mode === ListTileEnum.TILE ? 'flex-end' : 'inherit'} spacing={1}>
-				<Stack direction={'row'} spacing={1}>
-					<WebsiteIcon /> <a href={metadata.website}>{metadata.website}</a>
-				</Stack>
-				{itemContent.access === '0' ? (
-					<Stack direction={'row'} spacing={1}>
-						<LockIcon /> <Typography>Locked</Typography>
-					</Stack>
-				) : (
-					<Stack direction={'row'} spacing={1}>
-						<OpenLockIcon /> <Typography>Open</Typography>
-					</Stack>
-				)}
-				<Stack direction={'row'} spacing={1}>
-					<MemberIcon />
-					<Typography>{itemContent.memberCount || 0} Members</Typography>
-				</Stack>
-				{mode === ListTileEnum.LIST && <Box sx={{ flex: 1 }} />}
-				<Interactions />
-			</Stack>
-		)
-	}, [itemContent, metadata, mode])
-
-	if (!itemContent) return null
-
-	return mode === ListTileEnum.LIST ? (
-		<ListItem imageURL={imageURL} headline={itemContent.name} metaHeadline={bodyToText()} metaContent={metaContent}>
-			<Typography>{metadata.description}</Typography>
-		</ListItem>
-	) : (
-		<TileItem imageURL={imageURL} headline={itemContent.name} metaHeadline={bodyToText()} metaContent={metaContent}>
-			<Typography>{metadata.description}</Typography>
-		</TileItem>
-	)
-}
-
-const ItemList = (props) => {
-	const { content } = props
-
-	///
-	const [page, setPage] = React.useState(0)
-	const [rowsPerPage, setRowsPerPage] = React.useState(10)
-
-	const handleChangePage = (event, newPage) => {
-		setPage(newPage)
-	}
-
-	const handleChangeRowsPerPage = (event) => {
-		setRowsPerPage(+event.target.value)
-		setPage(0)
-	}
-	///
-
-	const [activePage, setActivePage] = useState(1)
-	const [totalPages, setTotalPages] = useState(0)
-	const [offset, setOffset] = useState(0)
-	const [itemsPerPage, setItemsPerPage] = useState(3)
-	const [displayMode, setDisplayMode] = useState(ListTileEnum.LIST)
-
-	const iPP = [3, 5, 9, 18, 36, 72]
-	const handleShowMoreItems = () => setItemsPerPage(itemsPerPage < iPP.length - 1 ? itemsPerPage + 1 : itemsPerPage)
-	const handleShowLessItems = () => setItemsPerPage(itemsPerPage > 0 ? itemsPerPage - 1 : 0)
-
-	useEffect(() => {
-		const _totalPages = Math.ceil(content.length / iPP[itemsPerPage])
-		setTotalPages(_totalPages)
-		if (activePage > _totalPages) setActivePage(_totalPages)
-	}, [content, itemsPerPage, activePage])
-
-	const handlePaginationChange = (e, { activePage }) => {
-		setActivePage(activePage)
-		setOffset((activePage - 1) * iPP[itemsPerPage])
-	}
-
-	const Wrapper = React.useMemo(() => (displayMode === ListTileEnum.LIST ? ListWrapper : TileWrapper), [displayMode])
-
-	if (!content) return null
-
-	// console.log(activePage,totalPages,offset,itemsPerPage)
-
-	return (
-		<Box>
-			<ListTileSwitch mode={displayMode} onSwitch={setDisplayMode} />
-			<Wrapper>
-				{content.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((d, i) => {
-					const _content = {
-						...d,
-					}
-					return <Item mode={displayMode} key={offset + i} content={_content} />
-				})}
-			</Wrapper>
-			<Box sx={{ my: 2 }} />
-			<TablePagination
-				rowsPerPageOptions={[10, 25, 100]}
-				component="div"
-				count={content.length}
-				rowsPerPage={rowsPerPage}
-				page={page}
-				onPageChange={handleChangePage}
-				onRowsPerPageChange={handleChangeRowsPerPage}
-			/>
-		</Box>
-	)
-}
 
 export const Component = (props) => {
 	const { api } = useSubstrate()
@@ -543,7 +292,7 @@ export const Component = (props) => {
 
 	return (
 		<Container maxWidth="lg">
-			<Typography component="h1" variant="h3">
+{/*			<Typography component="h1" variant="h3">
 				Organisations
 			</Typography>
 			<Stack direction="row" justifyContent="space-between" alignItems="center" spacing={12}>
@@ -558,7 +307,7 @@ export const Component = (props) => {
 					: <Button variant="outlined" startIcon={<AddIcon />} onClick={handleCreateBtn}> New DAO </Button>
 				}</Box>
 			</Stack>
-			<br />
+			<br />*/}
 		</Container>
 	)
 }
